+++ src/taskAutom/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "7.17.13"
+__version__ = "7.18.1"
 __author__ = 'Lucas Aimaretto'