+++ changelog.md
@@ -1,5 +1,12 @@
 # Versions #
 
+## [7.18.1] - 2023-01-30
+
+- Variables obtained through `argparse` are set to `True/False` depending on the input being `yes/no`. So now the treatment is hanlde as a boolen when needed.
+- New functionality to bulk updload files through SCP/SFTP.
+    - For this a new `dataFile` structure is needed with the following minimum columns: `ip|ftpLocalFilename|ftpRemoteFilename`.
+- Code reordering.
+
 ## [7.17.13] - 2023-01-27
 
 - CRON config update.